Make API more comparable to astral API
We want to get more accurate results by using the astral library. To that end, I've
reworked the API to be more similar to astral.

This has been done by pulling out the time calculations from Zmanim and moving them
to the location object.

Two notes that have come up:

  1. Should we move diaspora back from Location to Zmanim?
  2. The current implementation causes all the calculations to be done every time we
     try accessing a single item from the zmanim dictionary. A better way might be to
     create the calculations of sunrise, sunset etc as a separate property, and have
     zmanim dictionary only access the necessary calculations.